refactor: similar code usage being used in same files
